requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
Pillow
pytesseract
openai
langchain
langchain-community
langchain-openai
langchain-core
# chromadb 
pypdf